Fix Bugs and Code Optimization
